pyproject.toml
[project]
name = "tuberag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.4",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "youtube-transcript-api>=1.2.3",
]
